Refactor RPC calls into Channel.call with reflection Introduces one more copy from the unmarshalled message.
The common pattern for each client method was to send the request
message, then switch on the response, requiring each client method to
shutdown the channel on error or handle when the RPC chan was closed
from a connection closure.  This unifies the error handling and unifies
the handling of the synchronous (request/response) methods.