Rework shutdown logic for connections and channels Closes #29 Closes #23 Fixes #13
This reduces the number of places the connection and channel could be
shut down to a reader/writer problem, an async connection/channel error, or
synchronous connection/channel error.

A connection `call` method was also introduced to consolidate the error
handling in a similar way as the channel to clarify the channel open
handshakes.

The Error type replaces the Closed type and a smaller set of pre-defined
errors are allocated.  This type is used as return values for all
methods as well as notify channels.

An artificial protocolHeader frame is used to consolidate the error and
shutdown from the initial connection establishment.

Connection demuxing and opening took a large refactoring.

Channel and Connection locks are reduced, instead on the Channel the
`send` method is replaced with one that allows close events through but
no other, this captures the correct behavior during the `closing` and
`closed` states.

The `state` field is removed, using `sync.Once` instead to ensure the
destructor is only executed once.

The channel increment is now atomic, removing a lock.

Framing errors are now encapsulated in the Error type instead of being
pre-defined.