Propagate the channel.close method to the RPC calls
Deliver close message only when this is a synchronous response
(exception) to a waiting RPC call.  This is a touch racy when the call
goroutine is moving from send to receive.  If it doesn't meet the race,
the callee will get an ErrClosed instead of the text from this method.