Handle basic.ack{multiple: true} for publish confirms
RabbitMQ will attempt to coalesce acknowledgments and send a single ack
method to cover multiple delivery tags.  Responding properly requires
maintaining ordering of the acks produced and already consumed.

For example given 5 publishings, only 3 acks may arrive in as:

  2, 5, 4 (multiple)

To simplify the interface for ack listeners, the channels will only
receive individual delivery tags.  In such an example, the following
delivery tags will be send to the listener channels:

  2, 5, 1, 3, 4

Using individual acks instead of multiple simplifies receivers,
listeners can use a unordered data type like a map[uint64]bool to
maintain their pending state.

A heap is used instead of a simple slice for the delivery tags because
we the number of pending publishings could be in the millions.

Note: this implementation assumes at most once delivery semantics for
the acks.  The library will panic if a duplicate ack is received.  This
is intentional until the delivery semantics of publisher acks can be
determined (and relaxed in the library).