Isolate asynchronous channel close exceptions
Fixes #43

When an asynchronous method triggers a channel.close exception and the
client sends a synchronous channel.close method, the channel.close-ok
from the synchronous method may appear before the server exception
arrives.  When this happens the channel is closed and removed from the
dipatch pool so when the channel.close from the asynchronous exception
arrives, it was not able to be routed and caused a connection exception
from the client.

Fixed by handling channel.close and channel.close-ok in the connection
for missing channels per spec section 2.3.7.