After shutdown, close all future notify channels
To avoid non-obvious behavior around notify chan registration and race
conditions when a notify handler chan is added during shutdown, make
sure that the chan is closed during and after shutdown for all
notifications that will never be delivered.

Unifies struct field state mutex to be 'm' and the send mutex to be
'sendM' across Connection and Channel.

Closes #38