Isolate locks for channel registry
In response for finding a race that requires finer grained locking, use
a synchronized registry that manages the sequence and sequence to
channel map.

This removes the routability of a connection to channels during a
channel shutdown.  Assumes that no AMQP messages should arrive after a
connection shutdown arrives which may not be correct.