Anonymous queue binding in simpler-consumer
Queues declared with an empty string are anonymous queues and the server
will choose a queue name.

In order to bind anonymous queues, use the server chosen queue name for
the QueueBind call.

If an empty queue name is used in QueueBind, it looks like the server
will use the name of last queue declared on the channel.

* Use quoted strings for debug out.
* Fixes #66