Remove kill and verbose
Both the go test suite and travis will terminate long running tests.
This deferred kill makes an assumption on the execution isolation which
may not hold for future travis versions.

Verbose tests make it difficult to identify the failures due to protocol
logging.  Now that the tests have stabilized, only log the failures.