Fix data race when closing amqp connection
When using conn.NotifyClose, this later append to a list of channels
``me.closes``, while conn.shutdown() range over this list, so when a
connection is closed, ``conn.NotifyClose()`` may detect the connection closing
in the same time as the ``conn.shutdown()`` call, this is done in two
goroutine, which end up in data race where one appending to me.closes
and the other looping over it.

The fix is to move setting ``noNotify`` to True to the top of
``conn.shutdown()``, this way ``conn.NotifyClose()`` will not append to the
list anymore.