Use a separate mutex for publisher confirms state (see below)
Per discussion with @streadway.

Note that we intentionally do not synchronize access
for *reads* of this value because reads dramatically
outnumber writes in this case: publisher confirms
are typically enabled once but publishing on a channel can happen
100s of millions of times over the course of process life.
The contention is probably not worth atomically fetched
publisher confirm state reads.

Should we change our mind on the above, locking around me.confirming
reads can be trivially added.