Treat connection tests as integration tests
It's a new file, so add the licence header.

It's an integration test, so add the build tag.  `go test -tag
integration` to run this class of tests.  Alternatively we can consider
removing the build tags and skip tests if the AMQP_URL environment is
not set.

It's a test to use t.Fatalf instead of log.Fatalf in order to run other
tests.

Split acceptance critera in TestConcurrentClose independendently.

net.*op.Error looks like latent bug, so I called it out for more
investigation.