export default dial settings
When DialConfig() is called, no default value is set for the
Config.heartbeat and Config.Locale settings when they are 0.

Export the Dialer default values, so users can create custom DialConfigs
that use some of the default values without having to redefine and copy
them.
The DefaultConnectionTimeout is also exported to allow reusing the
default in custom dial functions